asasdasd

+++ chess.cpp
@@ -5,10 +5,10 @@
 
 
 
-	int Chessman::getSecondCoordinate() {
+	short Chessman::getSecondCoordinate() {
 		return this->_y;
 	}
-	int Chessman::getFirstCoordinate() {
+	short Chessman::getFirstCoordinate() {
 		return this->_x;
 	}
 	void Chessman::setSecondCoordinate(int num) {
@@ -26,7 +26,7 @@
 		else
 			this->_color = color;
 	}
-	int Chessman::getColor() { return this->_color; }
+	short Chessman::getColor() { return this->_color; }
 	void Chessman::move(int firstCor, int secondCor) {}
 	bool Chessman::isOutOfDesk(int firstCor, int secondCor) {
 		if (((firstCor < 1) || (firstCor > 8)) && ((secondCor < 1) || (secondCor > 8)))
@@ -40,7 +40,6 @@
 		else
 			return false;
 	}
-	
 	std::string Pawn::nameToString() {
 		return "I'm pawn.";
 	}
